sql/row: split up the writers
The file was quite large and a annoying to navigate.
Just split out Inserter, Updater and Deleter.

Release note: None
